@@ -1,3 +1,10 @@
-module pwd_manager
+module github.com/Ovec/pwd_manager
 
-go 1.18
+go 1.20
+
+require (
+	golang.org/x/crypto v0.16.0
+	golang.org/x/term v0.15.0
+)
+
+require golang.org/x/sys v0.15.0 // indirect
